refactor(storefront/api): move error handling logic
